pyproject.toml
[project]
name = "gravaity-agent"
version = "0.0.1"
description = "An agent with a tool to save long-term memory."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.0,<1.1.0",
    "langgraph-sdk>=0.2.2,<0.3.0",
    "langgraph-checkpoint>=2.0.0,<3.0.0",
    # LLM integrations
    "langchain>=1.0.0,<1.2.0",
    "langchain-core>=1.0.0,<1.2.0",
    "langchain-openai>=1.0.0,<1.2.0",
    "langchain-anthropic>=1.0.0,<1.3.0",
    "langchain-community>=0.4.0,<0.5.0",
    "python-dotenv>=1.0.1",
    "langsmith>=0.3.45",
    # RAG dependencies
    "langchain-postgres>=0.0.12,<0.1.0",
    "langchain-text-splitters>=0.3.11,<1.1.0",
    "pypdf>=5.1.0",
    "psycopg[binary]>=3.1.0",
    # API & streaming
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn>=0.30.0,<0.31.0",
    "redis>=5.0.0",
    # Document conversion
    "markitdown[all]>=0.0.1",
    "dashscope>=1.25.3"
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.1.71",
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent", "tools", "utils", "config", "db", "api", "infra"]
[tool.setuptools.package-dir]
"agent" = "src/agent"
"tools" = "src/tools"
"utils" = "src/utils"
"config" = "src/config"
"db" = "src/db"
"api" = "src/api"
"infra" = "src/infra"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_errors = true
